chore: move `server` to `api`
