+++ apps/api/main.py
@@ -1,4 +1,4 @@
-import server
+import api.server as server
 
 if __name__ == "__main__":
     server.run()
